% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_lon.R
\name{change_lon}
\alias{change_lon}
\title{Transform a raster for plotting}
\usage{
change_lon(x)
}
\arguments{
\item{rast}{The raster to be transformed}

\item{plot_extent}{plot extent. Maybe the extent of a raster to be plotted?}

\item{proj}{the projection the plot will be in}
}
\value{
A new raster that has been reprojected
}
\description{
This function changes the coordinate extent of a raster from (extent 0, 360)
to (extent -180, 180).
}
\note{
In development
}
\author{
Kimberlee Baldry
}
