% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get package data}
\usage{
get_data(type = "IceShelf")
}
\arguments{
\item{type}{IceShelf or CoastLines}
}
\value{
A shape object
}
\description{
This function returns data for ice shelves and coastlines
}
\note{
In development
}
\author{
Kimberlee Baldry
}
