% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot_proj.R
\name{map_plot_proj}
\alias{map_plot_proj}
\title{Create a boarder and coastline polygons for a plot}
\usage{
map_plot_proj(extent, proj_str, resolution, scale_res = 100, circle = F)
}
\arguments{
\item{extent}{plot extent in lat, lon coordinates. Maybe the extent of a raster to be plotted?}

\item{proj_str}{the projection the plot will be in}

\item{resolution}{the resolution of the data to be plotted}

\item{scale_res}{a scalar to help woth increasing the boarder resolution}

\item{circle}{is the boarder a circle?}
}
\value{
Two data frames that coresponds to the plot boarder and coastlines
}
\description{
This function returns data frames for a plot
boarder and coastlined that can be used in ggplot, based on a plot extent and projection.
This works for stereographic and curved segment plots
}
\note{
In development
}
\author{
Kimberlee Baldry
}
