% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_rast.R
\name{transform_rast}
\alias{transform_rast}
\title{Transform a raster for plotting}
\usage{
transform_rast(rast, proj, plot_extent)
}
\arguments{
\item{rast}{The raster to be transformed}

\item{proj}{the projection the plot will be in}

\item{plot_extent}{plot extent. Maybe the extent of a raster to be plotted?}
}
\value{
A new raster that has been reprojected
}
\description{
This function transforms a raster for plotting in a new projection space.
}
\note{
In development
}
\author{
Kimberlee Baldry
}
